replace v1 with v2
