+++ store/src/types/mod.rs
@@ -1,8 +1,4 @@
-pub mod prefix_v2;
-pub mod query_v2;
-// pub mod kv;
-pub mod kv_2;
-// pub mod multi;
-pub mod multi_v2;
-// pub mod prefix;
-// pub mod query;
+pub mod kv;
+pub mod multi;
+pub mod prefix;
+pub mod query;
