+++ store/src/types/prefix/mutable.rs
@@ -5,7 +5,6 @@ use crate::{
     WritePrefixStore,
 };
 
-/// Wraps an mutable KVStore with a prefix
 pub struct MutablePrefixStore<'a, DB> {
     pub(crate) store: KVStoreMut<'a, DB>,
     pub(crate) prefix: Vec<u8>,
