+++ store/src/types/query/versioned.rs
@@ -4,21 +4,21 @@ use database::{Database, PrefixDB};
 
 use crate::{
     error::KEY_EXISTS_MSG,
-    types::kv_2::immutable::{KVStoreBackend, KVStoreV2},
-    QueryableMultiKVStoreV2, StoreKey,
+    types::kv::immutable::{KVStore, KVStoreBackend},
+    QueryableMultiKVStore, StoreKey,
 };
 
-use super::kv::QueryKVStoreV2;
+use super::kv::QueryKVStore;
 
 pub struct VersionedQueryMultiStore<'a, DB, SK>(
-    pub(super) HashMap<&'a SK, QueryKVStoreV2<'a, PrefixDB<DB>>>,
+    pub(super) HashMap<&'a SK, QueryKVStore<'a, PrefixDB<DB>>>,
 );
 
-impl<DB: Database, SK: StoreKey> QueryableMultiKVStoreV2<PrefixDB<DB>, SK>
+impl<DB: Database, SK: StoreKey> QueryableMultiKVStore<PrefixDB<DB>, SK>
     for VersionedQueryMultiStore<'_, DB, SK>
 {
-    fn kv_store(&self, store_key: &SK) -> KVStoreV2<'_, PrefixDB<DB>> {
-        KVStoreV2(KVStoreBackend::Query(
+    fn kv_store(&self, store_key: &SK) -> KVStore<'_, PrefixDB<DB>> {
+        KVStore(KVStoreBackend::Query(
             self.0.get(store_key).expect(KEY_EXISTS_MSG),
         ))
     }
